wsl2_defconfig: Support up to 256 CPUs
Recent 64-bit consumer versions of Windows can support up to 256 CPUs.
The maximum setting of 8 CPUs in this kernel is an artificial limit;
developers from the WSL 2 team have said that the kernel should be able
to leverage all the CPUs available on the host on one socket.

Bump the limit so that users with more than 8 logical cores on their CPU
(e.g. Ryzen) can leverage all of their computing horsepower inside WSL
2.

Test: compare nproc output before & after on a Windows VM with faked
CPU topology
Signed-off-by: Danny Lin <danny@kdrag0n.dev>